Refactor to remove exec()
I used exec(), naively and without awareness of the potential security ramifications. It made my code so much more modular and easier to handle. I thought I'd found a perfect solution! For readability, perhaps. But at far too great a cost to the security of anyone else who decides to use my program.

I have put in a lot of effort to remove exec() from the code. It added about 100 extra lines to achieve the same thing as exec() did. But for removing such a glaring security risk, worth it.

Also:
- Moved function 'allOldEntriesGet' into getModifyValues.py
- Moved deDuplicateDBs.py into the root folder